Move the TypeScriptSemanticsIdentifier enum alongside the rest of the implementation
