+++ packages/open-schema-type-script/src/typeScriptSemanticsIdentifier.ts
